requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
pymysql==1.1.0
cryptography==41.0.7
Werkzeug==3.0.1
python-dotenv==1.0.0
email-validator==2.1.0
boto3==1.34.0
